requirements.txt
# Flask框架
Flask>=2.0.0
flask-cors>=3.0.0
PyJWT>=2.0.0

# 数据库
mysql-connector-python>=8.0.0
pymysql>=1.0.0

# 深度学习
torch>=1.10.0
torchvision>=0.11.0
clip @ git+https://github.com/openai/CLIP.git

# 数据处理
numpy>=1.20.0
Pillow>=8.0.0
ftfy>=6.0.0
regex>=2021.0.0
tqdm>=4.60.0

# 工具库
python-dotenv>=0.19.0
werkzeug>=2.0.0
